// front/src/generated/blame/api.ts
export interface Code {
    /**
     * 
     * @type {string}
     * @memberof Code
     */
    snippet?: string;
    /**
     * 
     * @type {string}
     * @memberof Code
     */
    link?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Code
     */
    highlight?: boolean;
}

/**
 * 
 * @export
 * @interface Report
 */
export interface Report {
    /**
     * 
     * @type {string}
     * @memberof Report
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof Report
     */
    search_url?: string;
    /**
     * 
     * @type {string}
     * @memberof Report
     */
    primary_link?: string;
    /**
     * 
     * @type {number}
     * @memberof Report
     */
    line_start_at?: number;
    /**
     * 
     * @type {number}
     * @memberof Report
     */
    total_ratio?: number;
    /**
     * 
     * @type {number}
     * @memberof Report
     */
    immediate_ratio?: number;
    /**
     * 
     * @type {string}
     * @memberof Report
     */
    label?: string;
    /**
     * 
     * @type {Array&lt;Code&gt;}
     * @memberof Report
     */
    code?: Array<Code>;
    /**
     * 
     * @type {Array&lt;Report&gt;}
     * @memberof Report
     */
    children?: Array<Report>;
}